team chat backend/requirements.txt
fastapi
uvicorn[standard]

# Database
sqlalchemy
psycopg2-binary

# Auth & Security
python-jose[cryptography]
passlib[bcrypt]

# Environment variables
python-dotenv

# Validation
pydantic
